fix(feature/coredata): fixed last commit
